add to wishlist test convert to junit
